jitsy/jitsi-recorder-lite/requirements.txt
# Core dependencies for WebRTC Bot
websockets==12.0
aiortc==1.9.*
av==12.0.*
aiohttp==3.9.*

# Optional: S3 upload support
boto3==1.34.*

# Optional: Redis coordination (for future Prosody integration)
redis==5.0.*

# Utils
python-dotenv==1.0.*
